+++ iotdb-core/metrics/core/src/main/java/org/apache/iotdb/metrics/core/type/IoTDBTimerHistogramSnapshot.java
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-package org.apache.iotdb.metrics.micrometer.type;
+package org.apache.iotdb.metrics.core.type;
 
 import org.apache.iotdb.metrics.type.HistogramSnapshot;
